# bin/Release/net7.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "shoppingWebApi, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/Auth/Register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UserRegisterDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Int32ServiceResponse"
            }
          }
        }
      }
    },
    "/Auth/Login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UserLoginDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Int32ServiceResponse"
            }
          }
        }
      }
    },
    "/api/Order/GetAll": {
      "get": {
        "tags": [
          "Order"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/GetOrderDtoListServiceResponse"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "oauth2": [ ]
          }
        ]
      }
    },
    "/api/Order/item": {
      "post": {
        "tags": [
          "Order"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/AddProductItemDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/GetProductItemDtoServiceResponse"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "oauth2": [ ]
          }
        ]
      }
    },
    "/api/Order/{id}": {
      "get": {
        "tags": [
          "Order"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/GetOrderDtoServiceResponse"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "oauth2": [ ]
          }
        ]
      },
      "delete": {
        "tags": [
          "Order"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/GetOrderDtoServiceResponse"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "oauth2": [ ]
          }
        ]
      }
    },
    "/api/Order": {
      "put": {
        "tags": [
          "Order"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/GetOrderDtoListServiceResponse"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "oauth2": [ ]
          }
        ]
      }
    },
    "/api/Product": {
      "get": {
        "tags": [
          "Product"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/GetProductDtoListServiceResponse"
            }
          }
        }
      }
    },
    "/WeatherForecast": {
      "get": {
        "tags": [
          "WeatherForecast"
        ],
        "operationId": "GetWeatherForecast",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/WeatherForecast"
              }
            }
          }
        }
      }
    }
  },
  "definitions": {
    "AddProductItemDto": {
      "type": "object",
      "properties": {
        "productId": {
          "format": "int32",
          "type": "integer"
        },
        "number": {
          "format": "int32",
          "type": "integer"
        }
      },
      "additionalProperties": false
    },
    "GetOrderDto": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "item": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/GetProductItemDto"
          }
        },
        "status": {
          "$ref": "#/definitions/OrderStatus"
        }
      },
      "additionalProperties": false
    },
    "GetOrderDtoListServiceResponse": {
      "type": "object",
      "properties": {
        "data": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/GetOrderDto"
          }
        },
        "success": {
          "type": "boolean"
        },
        "message": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "GetOrderDtoServiceResponse": {
      "type": "object",
      "properties": {
        "data": {
          "$ref": "#/definitions/GetOrderDto"
        },
        "success": {
          "type": "boolean"
        },
        "message": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "GetProductDto": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "brand": {
          "type": "string"
        },
        "model": {
          "type": "string"
        },
        "type": {
          "type": "string"
        },
        "price": {
          "format": "int32",
          "type": "integer"
        },
        "rating": {
          "format": "int32",
          "type": "integer"
        },
        "stock": {
          "format": "int32",
          "type": "integer"
        },
        "waterproof": {
          "type": "boolean"
        },
        "fastCharge": {
          "type": "boolean"
        },
        "image": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "GetProductDtoListServiceResponse": {
      "type": "object",
      "properties": {
        "data": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/GetProductDto"
          }
        },
        "success": {
          "type": "boolean"
        },
        "message": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "GetProductItemDto": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "product": {
          "$ref": "#/definitions/GetProductDto"
        },
        "number": {
          "format": "int32",
          "type": "integer"
        }
      },
      "additionalProperties": false
    },
    "GetProductItemDtoServiceResponse": {
      "type": "object",
      "properties": {
        "data": {
          "$ref": "#/definitions/GetProductItemDto"
        },
        "success": {
          "type": "boolean"
        },
        "message": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "Int32ServiceResponse": {
      "type": "object",
      "properties": {
        "data": {
          "format": "int32",
          "type": "integer"
        },
        "success": {
          "type": "boolean"
        },
        "message": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "OrderStatus": {
      "enum": [
        "Cart",
        "Ordered"
      ],
      "type": "string"
    },
    "UserLoginDto": {
      "type": "object",
      "properties": {
        "username": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "UserRegisterDto": {
      "type": "object",
      "properties": {
        "username": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "WeatherForecast": {
      "type": "object",
      "properties": {
        "date": {
          "format": "date",
          "type": "string"
        },
        "temperatureC": {
          "format": "int32",
          "type": "integer"
        },
        "temperatureF": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        },
        "summary": {
          "type": "string"
        }
      },
      "additionalProperties": false
    }
  },
  "securityDefinitions": {
    "oauth2": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Standard Authorization header using the Bearer scheme. Example: \"bearer {token}\" "
    }
  }
}